补充功能

+++ src/main/java/com/quinit/aaaa/sevice/impl/DeptServiceImpl.java
@@ -1,12 +1,13 @@
 package com.quinit.aaaa.sevice.impl;
 
+import com.quinit.aaaa.exception.DeptNotEmptyException;
 import com.quinit.aaaa.mapper.DeptMapper;
+import com.quinit.aaaa.mapper.EmpMapper;
 import com.quinit.aaaa.pojo.Dept;
 import com.quinit.aaaa.sevice.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.sql.Time;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -15,14 +16,21 @@ public class DeptServiceImpl implements DeptService {
 
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
     @Override
     public List<Dept> findAll() {
         return deptMapper.findAll();
     }
 
     @Override
-    public void deleteById(Integer id) {
-        deptMapper.deleteById(id);
+    public void deleteById(Integer id) throws DeptNotEmptyException {
+        if(empMapper.deptEmpCount(id)==0){
+            deptMapper.deleteById(id);
+        }
+        else{
+            throw new DeptNotEmptyException("部门下有员工，不能删除");
+        };
     }
 
     @Override
